requirements.txt
torch
pyro-ppl
matplotlib
numpy
scipy
